Updated Code to match UI
Updated:
US 01.01.01 As an entrant, I want to join the waiting list for a specific event
US 01.01.02 As an entrant, I want to leave the waiting list for a specific event

US 03.07.01 As an administrator I want to remove organizers that violate app policy.
US 03.02.01 As an administrator, I want to be able to remove profiles.
US 03.05.01 As an administrator, I want to be able to browse profiles.